+++ src/domain/mod.rs
@@ -13,6 +13,7 @@ pub mod detail;
 pub mod device_detail;
 pub mod device_view;
 pub mod interface_view;
+pub mod ip_range_view;
 pub mod ip_view;
 pub mod prefix_view;
 pub mod rack_view;